renamed module catalog_handlers_utils -> catalog_utils
