+++ services/web/server/src/simcore_service_webserver/catalog_handlers.py
@@ -12,7 +12,6 @@ from pydantic import ValidationError
 from . import catalog_client
 from ._constants import RQ_PRODUCT_KEY
 from ._meta import api_version_prefix
-from .catalog_handlers_utils import can_connect
 from .catalog_models import (
     ServiceInputGet,
     ServiceInputKey,
@@ -23,6 +22,7 @@ from .catalog_models import (
     json_dumps,
     replace_service_input_outputs,
 )
+from .catalog_utils import can_connect
 from .login.decorators import RQT_USERID_KEY, login_required
 from .rest_constants import RESPONSE_MODEL_POLICY
 from .security_decorators import permission_required
