+++ services/web/server/src/simcore_service_webserver/catalog_utils.py
